pyproject.toml
[project]
name = "devops-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.16",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "azure-common==1.1.28",
    "azure-core==1.33.0",
    "azure-identity==1.21.0",
    "azure-mgmt-compute==34.1.0",
    "azure-mgmt-core==1.5.0",
    "azure-mgmt-network==28.1.0",
    "azure-mgmt-resource==23.3.0",
    "azure-mgmt-storage==22.1.1",
    "beautifulsoup4==4.13.3",
    "boto3==1.37.33",
    "botocore==1.37.33",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "cryptography==44.0.2",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "fastapi==0.115.12",
    "frozenlist==1.5.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google==3.0.0",
    "google-api-core==2.24.2",
    "google-auth==2.38.0",
    "google-cloud==0.34.0",
    "google-cloud-compute==1.29.0",
    "google-cloud-resource-manager==1.14.2",
    "googleapis-common-protos==1.70.0",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "idna==3.10",
    "isodate==0.7.2",
    "jiter==0.9.0",
    "jmespath==1.0.1",
    "msal==1.32.0",
    "msal-extensions==1.3.1",
    "multidict==6.4.3",
    "openai==1.74.0",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pygithub==2.6.1",
    "pyjwt==2.10.1",
    "pynacl==1.5.0",
    "python-dateutil==2.9.0.post0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rsa==4.9",
    "s3transfer==0.11.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "starlette==0.46.2",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.1",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.19.0",
]
